An example of using contracts attributes.
This adds a new version of the shopping cart example (and removes the old one) which is richer and uses attributes from the contracts crate. The example is designed such contracts are executed (and tested) at runtime thus validating the soundness of the MIRAI assertions to the contracts crate.

Also attempted to verify this example using MIRAI and identified some gaps. See also TODO's in src/main.rs.