Enable contracts in integration tests (#347)
* Enable contracts in integration tests

This enables use of contract annotations in integration tests. The integration test driver is extended to this end to deal with multiple --extern libraries.

A new test case is added based on contracts but this doesn't do much yet because of missing summaries.

* Treat linux/windows external libraries

* Document problem with double assume and fix it in a better way

* Add a repro test for failure in inferring closure preconditions

* Fixing --extern discovery for windows

Also adding copyright headers and eliminating unused warnings.