Fixing contracts dependency
I broke the contracts dependencies by re-creating my fork. This updates the dependencies so they work again after the re-creation.

The breakage was also because I used a revision based reference to the repo, for stability for CLion run configurations which require to refer to names in target/debug. This is now again symbolic, and should be anway have enough stability via Cargo.lock.