Enabling analysis of tests, plus turning angelic treatment off.
This adds two new options `--test_only` and `--diag=<level>` which allow for
analyzing test functions as well as turning off the transient propagation
of angelic treatment in the presence of functions without a body. This
flags are intended to use together, but are technically independent.

`--test_only` will cause MIRAI to look for a test runner and only
analyze those methods which are called by the test runner (i.e.
those which have the `#[test]` attribute), plus methods in their
call graph. The implementation of this isn't as trivial as
expected since the rust test runner uses a native method and
builds data structures for the set of test methods, while the
original `#[test]` is removed from the source by macro expansion.
We use heuristics which result in a first implementation which may
need further refinement.

One must use the `--test` (rustc) or `--tests` (cargo) option in
addition to `TEST_ONLY=1` to get the tests analyzed.

`--diag=strict` will simply cause the analyzer to report with info log
that it can't analyze a method reliably, but continues instead
of bailing out.

A new test has been added and the test infrastructure extended to
be able to deal with those options (a marker in the source indicates
for integration tests what to do).