Make ./setup.sh more robust
The current setup script runs under the context of the directory from which it is called. If that's the directory in which "./setup.sh" lives, it will pickup the right 'rust-toolchain'. Otherwise it won't.

This fixes this fragility by fetching the rust-toolchain file contents and passing them as an explicit rustup argument.